+++ lib/ferrum/page.rb
@@ -388,8 +388,7 @@ module Ferrum
       when :request
         client.on("Fetch.requestPaused") do |params, index, total|
           request = Network::InterceptedRequest.new(client, params)
-          exchange = network.select(request.network_id).last
-          exchange ||= network.build_exchange(request.network_id)
+          exchange = network.find_or_build_exchange(request.network_id)
           exchange.intercepted_request = request
           block.call(request, index, total)
         end